style: lint the code & remove unused imports/ variables
