+++ src/components/FileCard.tsx
@@ -1,4 +1,5 @@
 import { Box, BoxProps, Flex, HStack } from '@chakra-ui/react';
+
 import { useConfigContext } from 'contexts/config';
 
 type FileCardProps = {
