+++ src/components/FolderCard.tsx
@@ -1,4 +1,5 @@
 import { Box, Flex, HStack } from '@chakra-ui/react';
+
 import { useConfigContext } from 'contexts/config';
 
 type FolderCardProps = {
