+++ src/components/ProfileCard.tsx
@@ -2,7 +2,6 @@ import { CopyIcon } from '@chakra-ui/icons';
 import { Box, Button, Flex, Text, Tooltip, VStack } from '@chakra-ui/react';
 
 import type { IPCContact } from 'types/types';
-
 import UpdateContact from './contact/UpdateContact';
 
 type ProfileCardProps = {
