+++ src/components/computing/CustomButtons.tsx
@@ -1,4 +1,5 @@
 import { Button, useColorModeValue } from '@chakra-ui/react';
+
 import { useConfigContext } from 'contexts/config';
 
 type CustomButtonProps = {
