+++ src/components/file/UploadFile.tsx
@@ -1,3 +1,4 @@
+import { ChangeEvent, useState } from 'react';
 import {
 	Button,
 	HStack,
@@ -5,37 +6,33 @@ import {
 	Input,
 	Text,
 	useBreakpointValue,
-	useColorModeValue,
 	useDisclosure,
 	useToast,
 } from '@chakra-ui/react';
-import { ChangeEvent, useState } from 'react';
+import { AiOutlineFileAdd } from 'react-icons/ai';
 
 import { v4 as uuid } from 'uuid';
 
 import Modal from 'components/Modal';
-import type { IPCFile } from 'types/types';
 
 import { extractFilename, getFileContent } from 'utils/fileManipulation';
 import generateFileKey from 'utils/generateFileKey';
 
-import { useConfigContext } from 'contexts/config';
 import { useDriveContext } from 'contexts/drive';
 import { useUserContext } from 'contexts/user';
-import { AiOutlineFileAdd } from 'react-icons/ai';
+
+import type { IPCFile } from 'types/types';
 
 const UploadFile = (): JSX.Element => {
 	const { user } = useUserContext();
-	const { config } = useConfigContext();
 	const { path, files, setFiles } = useDriveContext();
-	const toast = useToast({ duration: 2000, isClosable: true });
 
 	const [fileEvent, setFileEvent] = useState<ChangeEvent<HTMLInputElement> | undefined>(undefined);
 	const [isLoading, setIsLoading] = useState(false);
 	const { isOpen, onOpen, onClose } = useDisclosure();
-	const colorText = useColorModeValue('gray.800', 'white');
 
 	const isDrawer = useBreakpointValue({ base: true, sm: false }) || false;
+	const toast = useToast({ duration: 2000, isClosable: true });
 
 	const uploadFile = async () => {
 		if (!fileEvent) return;
